fixed JSON-Serde Class Not Found Exception and Updated images
For certain hive queries that created tables and referenced another
table to used the JSON-Serde library to work with Json data or queries
that loaded the table data, I needed to include the path to the
JSON-Serde library in order to avoid class not found exception.

I updated pictures to be up to date with latest HDP 2.6 sandbox image
features.