Created Acquiring HVAC Sensor Data tutorial
This tutorial module covers using NiFi to acquire sensor data over
HTTP, preprocessing the data and storing it into HDFS.
This tutorial offers 3 approaches in which a developer decides to auto
deploy the NiFi flow using NiFi Rest API Call, Manually import the NiFi
flow via NiFi UI or Manually build the NiFi data flow.

Progress: auto deploy of nifi flow script has been tested and works.
Manually importing the NiFi flow has been tested and works too.
Manually building the NiFi flow steps are included and needs to be
reviewed just to verify steps were not missed for people who are new to
nifi.

Note: Need to enrich the content of this tutorial cause right now it is
just a skeleton to build this portion of the HVAC demo.