Created nifi-auto-deploy.sh
Script is going to automate uploading, importing and starting nifi flow.

The big part of this script is leveraging a python script that parses
the XML nifi template and adds the KEYS and TOKENS.

More updates coming, but first need to finish updating the content of
the customer sentiment analysis tutorial.